Ambiguity around using a CNAME on Route53
Using a CNAME for an ALIAS on Route53 is not supported. I have aligned the wording to be similar to the Lightsail example and removed the note about CNAMEs.

In this case we are manually specifying the Alias Target and thus its possible to create a CNAME record with and ALIAS that won't resolve.